+++ packages/breadcrumb/src/component/Breadcrumb/Breadcrumb.config.js
@@ -1,0 +1,1 @@
+export const BREADCRUMB_KEY = 'breadcrumb';
