--- packages/product-breadcrumb/src/component/ProductBreadcrumb/index.js
@@ -1,1 +1,0 @@
-export { default, default as ProductBreadcrumb } from './ProductBreadcrumb.component';
